Add HTTPServer (#17)
Sometimes you need to mock HTTP servers where the only option is to provide a URL endpoint through to your test. That is, when you do not have direct access to the router, or it's impractical to inject the behavior. For case this you can use a `HTTPServer`.

Using a real HTTP server has some benefits:

1. It's isolated. That means it does not interfere in anyway with the global HTTP server.

2. It can be used in parallel. You can either share the same `HTTPServer` across many tests (such as in `TestMain`), or create one for each test in parallel. Providing `0` for the port (as in the example above) will ensure that it always selects an unused random port.

3. It mutable. After creating and starting the `HTTPServer` you can add/remove handlers. This is useful when most tests need a base logic, but some cases need to return special under specific scenarios.

You can create your own handlers, of course, but there are a few common ones that also ship with `tf`